move user_collection.py to db/
