+++ ext/structs/__init__.py
@@ -1,4 +1,4 @@
 from .proxies.dir_proxy import DirProxy
 from .proxies.dict_proxy import DictProxy
-from .user_collection import UserCollection
+from .db.user_collection import UserCollection
 from .simple import *
